requirements.txt
flask
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-genai
python-dotenv